Android: initial setup
Changes introduced by this commit don't let you build 'apk' file! This
commit consists of files from various libraries we are depended on in
their original form. They will be customized to our needs in the next
commit and only then you will be able to successfully build 'apk'. This
commit is considered as a reference point for future changes.