Android: use 'SDL_RWops' structure to load files from 'assets'
On Android we can use standard C or C++ functions to load a file from
sdcard or any location we know the full path to and have appropriate
permissions, but not from 'assets' directory since it is compressed
inside 'apk' archive. 'SDL_RWops' is a layer of abstraction over I/O
machinery of every system supported by SDL. In case of Android it does
all heavy lifting with loading data from 'assets' directory.